day-5/requirements.txt
# ==============================================================================
# Python Dependencies for Day 5: Battle Agent
# ==============================================================================

# Core CrewAI with tools
crewai[tools]>=0.86.0

# Environment management
python-dotenv>=1.0.0

# FastAPI for deployment
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Web search tools (choose one or more)
# google-search-results>=2.4.2  # Serper
# requests>=2.31.0

# Optional: Additional specialized tools
# yfinance>=0.2.0  # Stock data
# python-weather>=0.4.0  # Weather
# newsapi-python>=0.2.7  # News

